+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Split `EnhancedInputSystems::Apply` into two sets: `Apply` (where `Action` gets updated) and `Trigger` (where observers get triggered)
 
+### Fixed
+
+- Capture input that is pressed and released within a single frame.
+
 ## [0.26.0] - 2026-06-20
 
 ### Changed